+++ CHANGELOG.md
@@ -5,11 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [Unreleased]
+
+- N/A
+
+## [2.0.0] - 2021-11-14
 
 ### Added
 
-- Websocket support for updating story text and player voting.
+- Add Websocket support
+  - use websocket for updating story text and player votes and notifying other game participants about changes
+  - websocket server management of sessions and message broadcasting via `WebSocketServer` module; includes tests and mock
+  - move all game logic out of Cloudflare Worker and into the Cloudflare Durable Object
+  - configure websocket support for local development using miniflare and proxy
+  - add Cloudflare TypeScript typings for CloudflareWebSocket. Add `tsconfig.json` for VSCode support.
+  - authenticate user session/auth before setting up websocket connection.
+
+### Removed
+
+- Removed http polling. Remove http updating. Still use http for retrieving full game state.
 
 ## [1.0.1] - 2021-11-03
 
@@ -32,7 +46,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hosted on Cloudflare Pages
 - Changelog
 
-[unreleased]: https://github.com/briangershon/planning-poker/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/briangershon/planning-poker/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/briangershon/planning-poker/releases/tag/v2.0.0
 [1.0.1]: https://github.com/briangershon/planning-poker/releases/tag/v1.0.1
 [1.0.0]: https://github.com/briangershon/planning-poker/releases/tag/v1.0.0
 [0.0.1]: https://github.com/briangershon/planning-poker/releases/tag/v0.0.1